tests/requirements.txt
pytest
pytest-django
pytest-subtests